Add Delete Guardian Functionality to UI

+++ Destiny.ScrimTracker.Logic/Repositories/GuardianMatchResultsRepository.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Reflection;
 using Destiny.ScrimTracker.Logic.Adapters;
 using Destiny.ScrimTracker.Logic.Models;
+using Microsoft.EntityFrameworkCore.Internal;
 
 namespace Destiny.ScrimTracker.Logic.Repositories
 {
@@ -13,6 +14,7 @@ namespace Destiny.ScrimTracker.Logic.Repositories
         IEnumerable<GuardianMatchResult> GetGuardianMatchResults(string matchId, string teamId);
         IEnumerable<GuardianMatchResult> GetMatchResultsForGuardian(string guardianId);
         IEnumerable<string> DeleteGuardianResults(string matchId);
+        IEnumerable<string> DeleteAllResultsForGuardian(string guardianId);
     }
     
     public class GuardianMatchResultsRepository : IGuardianMatchResultsRepository
@@ -41,12 +43,34 @@ namespace Destiny.ScrimTracker.Logic.Repositories
 
         public IEnumerable<GuardianMatchResult> GetMatchResultsForGuardian(string guardianId)
         {
-            return _databaseContext.GuardianMatchResults.Where(res => res.GuardianId == guardianId);
+            var matchResults = _databaseContext.GuardianMatchResults.Where(res => res.GuardianId == guardianId);
+
+            if (matchResults == null || !matchResults.Any())
+            {
+                return new GuardianMatchResult[] {};
+            }
+
+            return matchResults;
         }
 
         public IEnumerable<string> DeleteGuardianResults(string matchId)
         {
             var results = _databaseContext.GuardianMatchResults.Where(gmr => gmr.MatchId == matchId);
+            
+            _databaseContext.RemoveRange(results);
+            _databaseContext.SaveChanges();
+
+            return results.Select(gmr => gmr.Id);
+        }
+
+        public IEnumerable<string> DeleteAllResultsForGuardian(string guardianId)
+        {
+            var results = _databaseContext.GuardianMatchResults.Where(gmr => gmr.GuardianId == guardianId);
+
+            if (results == null || !results.Any())
+            {
+                return null;
+            }
             
             _databaseContext.RemoveRange(results);
             _databaseContext.SaveChanges();
